+++ grails-app/controllers/de/laser/AdminController.groovy
@@ -804,6 +804,7 @@ class AdminController  {
     }
 
     @Secured(['ROLE_ADMIN'])
+    @Transactional
     def tippTransfer(){
         log.debug("tippTransfer :: ${params}")
         Map<String, Object> result = [:]
@@ -815,7 +816,7 @@ class AdminController  {
       if(ti && tipp){
         tipp.title = ti
         try{
-          tipp.save(flush:true,failOnError:true)
+          tipp.save(failOnError:true)
           result.success = true
         }catch(Exception e){
             log.error( e.toString() )
@@ -831,6 +832,7 @@ class AdminController  {
     }
 
     @Secured(['ROLE_ADMIN'])
+    @Transactional
     def ieTransfer(){
         log.debug( params.toMapString() )
         Map<String, Object> result = [:]
@@ -844,7 +846,7 @@ class AdminController  {
             def sourceIEs = IssueEntitlement.findAllByTipp(result.sourceTIPPObj)
             sourceIEs.each{
                 it.setTipp(result.targetTIPPObj)
-                it.save(flush:true)
+                it.save()
             }
         }
