+++ grails-app/services/com/k_int/kbplus/PendingChangeService.groovy
@@ -713,6 +713,8 @@ class PendingChangeService extends AbstractLockableService {
         ret
     }
 
-
+    void acknowledgeChange(PendingChange changeAccepted) {
+        changeAccepted.delete()
+    }
 
 }
